legacy/requirements.txt
# JetSki Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# AI & LLM
openai==1.3.5
google-genai>=0.2.0  # Google Gemini 2.5 Flash Image (NanoBanana)
langchain==0.0.339
langchain-openai==0.0.2
crewai==0.1.0

# Image Processing
Pillow>=10.0.0  # For image manipulation and saving

# YouTube & Transcription
youtube-transcript-api==0.6.1
pytube==15.0.0
yt-dlp==2023.11.16

# Google APIs (Docs & Drive)
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Database - Supabase for data persistence
supabase==2.9.0

# Utilities
pydantic==2.5.1
requests==2.31.0
python-multipart==0.0.6

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0

